updating and fixing and cleaning and adding comments

+++ backend/routes/linesmen.js
@@ -49,6 +49,9 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
   try {
     const { id } = req.params;
+    const linesman = await Linesman.findById(id);
+    if (!linesman) return res.status(404).send('Linesman not found.');
+
     const { error } = linesmanValidationSchema.validate(req.body);
     if (error) {
       return res.status(400).json({ message: error.details[0].message });
@@ -59,13 +62,15 @@ router.put('/:id', async (req, res) => {
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
-  
 });
 
 // Delete a Linesman
 router.delete('/:id', async (req, res) => {
   try {
     const { id } = req.params;
+    const linesman = await Linesman.findById(id);
+    if (!linesman) return res.status(404).send('Linesman not found.');
+
     await Linesman.findByIdAndDelete(id);
     res.json({ message: 'Linesman deleted successfully' });
   } catch (error) {
