+++ backend/routes/teams.js
@@ -49,6 +49,9 @@ router.post('/', async (req, res) => {
 router.put('/:id', async (req, res) => {
   try {
     const { id } = req.params;
+    const team = await Team.findById(id);
+    if (!team) return res.status(404).send('Team not found.');
+
     const { error } = teamValidationSchema.validate(req.body);
     if (error) {
       return res.status(400).json({ message: error.details[0].message });
@@ -59,13 +62,15 @@ router.put('/:id', async (req, res) => {
   } catch (error) {
     res.status(400).json({ message: error.message });
   }
-  
 });
 
 // Delete a Team
 router.delete('/:id', async (req, res) => {
   try {
     const { id } = req.params;
+    const team = await Team.findById(id);
+    if (!team) return res.status(404).send('Team not found.');
+
     await Team.findByIdAndDelete(id);
     res.json({ message: 'Team deleted successfully' });
   } catch (error) {
